improve: reduce copy-pastes
