+++ algebird-core/src/main/scala/com/twitter/algebird/Aggregator.scala
@@ -10,9 +10,7 @@ import scala.collection.generic.CanBuildFrom
  * To create a parallel aggregator that operates on a single input in parallel, use:
  * GeneratedTupleAggregator.from2((agg1, agg2))
  */
-object Aggregator extends java.io.Serializable {
-  implicit def applicative[I]: Applicative[({ type L[O] = Aggregator[I, _, O] })#L] =
-    new AggregatorApplicative[I]
+object Aggregator extends java.io.Serializable with AggregatorCompat {
 
   private val DefaultSeed = 471312384
 
@@ -20,7 +18,7 @@ object Aggregator extends java.io.Serializable {
    * This is a trivial aggregator that always returns a single value
    */
   def const[T](t: T): MonoidAggregator[Any, Unit, T] =
-    prepareMonoid { _: Any => () }.andThenPresent(_ => t)
+    prepareMonoid { (_: Any) => () }.andThenPresent(_ => t)
 
   /**
    * Using Aggregator.prepare,present you can add to this aggregator
@@ -172,7 +170,7 @@ object Aggregator extends java.io.Serializable {
    * How many items satisfy a predicate
    */
   def count[T](pred: T => Boolean): MonoidAggregator[T, Long, Long] =
-    prepareMonoid { t: T => if (pred(t)) 1L else 0L }
+    prepareMonoid { (t: T) => if (pred(t)) 1L else 0L }
 
   /**
    * Do any items satisfy some predicate
@@ -310,7 +308,7 @@ object Aggregator extends java.io.Serializable {
    * Put everything in a Set. Note, this could fill the memory if the Set is very large.
    */
   def toSet[T]: MonoidAggregator[T, Set[T], Set[T]] =
-    prepareMonoid { t: T => Set(t) }
+    prepareMonoid { (t: T) => Set(t) }
 
   /**
    * This builds an in-memory Set, and then finally gets the size of that set. This may not be scalable if the
@@ -506,41 +504,7 @@ trait Aggregator[-A, B, +C] extends java.io.Serializable { self =>
     }
 }
 
-/**
- * Aggregators are Applicatives, but this hides the middle type. If you need a join that does not hide the
- * middle type use join on the trait, or GeneratedTupleAggregator.fromN
- */
-class AggregatorApplicative[I] extends Applicative[({ type L[O] = Aggregator[I, _, O] })#L] {
-  override def map[T, U](mt: Aggregator[I, _, T])(fn: T => U): Aggregator[I, _, U] =
-    mt.andThenPresent(fn)
-  override def apply[T](v: T): Aggregator[I, _, T] =
-    Aggregator.const(v)
-  override def join[T, U](mt: Aggregator[I, _, T], mu: Aggregator[I, _, U]): Aggregator[I, _, (T, U)] =
-    mt.join(mu)
-  override def join[T1, T2, T3](
-      m1: Aggregator[I, _, T1],
-      m2: Aggregator[I, _, T2],
-      m3: Aggregator[I, _, T3]
-  ): Aggregator[I, _, (T1, T2, T3)] =
-    GeneratedTupleAggregator.from3((m1, m2, m3))
 
-  override def join[T1, T2, T3, T4](
-      m1: Aggregator[I, _, T1],
-      m2: Aggregator[I, _, T2],
-      m3: Aggregator[I, _, T3],
-      m4: Aggregator[I, _, T4]
-  ): Aggregator[I, _, (T1, T2, T3, T4)] =
-    GeneratedTupleAggregator.from4((m1, m2, m3, m4))
-
-  override def join[T1, T2, T3, T4, T5](
-      m1: Aggregator[I, _, T1],
-      m2: Aggregator[I, _, T2],
-      m3: Aggregator[I, _, T3],
-      m4: Aggregator[I, _, T4],
-      m5: Aggregator[I, _, T5]
-  ): Aggregator[I, _, (T1, T2, T3, T4, T5)] =
-    GeneratedTupleAggregator.from5((m1, m2, m3, m4, m5))
-}
 
 trait MonoidAggregator[-A, B, +C] extends Aggregator[A, B, C] { self =>
   def monoid: Monoid[B]
