+++ algebird-core/src/main/scala/com/twitter/algebird/Fold.scala
@@ -173,13 +173,9 @@ final class FoldState[X, -I, +O] private[algebird] (val add: (X, I) => X, val st
  * recommended that "end" functions not mutate the accumulator in order to support scans (producing a stream
  * of intermediate outputs by calling "end" at each step).
  */
-object Fold extends CompatFold {
+object Fold extends CompatFold with FoldApplicativeCompat {
 
-  /**
-   * "import Fold.applicative" will bring the Applicative instance into scope. See FoldApplicative.
-   */
-  implicit def applicative[I]: Applicative[Fold[I, _]] =
-    new FoldApplicative[I]
+
 
   /**
    * Turn a common Scala foldLeft into a Fold. The accumulator MUST be immutable and serializable.
@@ -335,18 +331,4 @@ object Fold extends CompatFold {
     }
 }
 
-/**
- * Folds are Applicatives!
- */
-class FoldApplicative[I] extends Applicative[Fold[I, _]] {
-  override def map[T, U](mt: Fold[I, T])(fn: T => U): Fold[I, U] =
-    mt.map(fn)
-  override def apply[T](v: T): Fold[I, T] =
-    Fold.const(v)
-  override def join[T, U](mt: Fold[I, T], mu: Fold[I, U]): Fold[I, (T, U)] =
-    mt.join(mu)
-  override def sequence[T](ms: Seq[Fold[I, T]]): Fold[I, Seq[T]] =
-    Fold.sequence(ms)
-  override def joinWith[T, U, V](mt: Fold[I, T], mu: Fold[I, U])(fn: (T, U) => V): Fold[I, V] =
-    mt.joinWith(mu)(fn)
-}
+
